move `recover_evaluations_in_domain_order` into cosets.rs
